+++ leia-core/src/main/java/com/grookage/leia/core/retrieval/RepositoryRefresher.java
@@ -24,8 +24,8 @@ public class RepositoryRefresher extends AbstractLeiaRefresher<SchemaRegistry> {
 
     @Builder
     protected RepositoryRefresher(final RepositorySupplier supplier,
-                                  final int configRefreshTimeSeconds) {
-        super(supplier, configRefreshTimeSeconds);
+                                  final int refreshTimeInSeconds) {
+        super(supplier, refreshTimeInSeconds);
     }
 
 }