+++ leia-schema-validator/src/test/java/com/grookage/leia/validator/SchemaValidationUtilsTest.java
@@ -17,26 +17,40 @@
 package com.grookage.leia.validator;
 
 import com.grookage.leia.models.ResourceHelper;
-import com.grookage.leia.models.attributes.*;
+import com.grookage.leia.models.attributes.ArrayAttribute;
+import com.grookage.leia.models.attributes.BooleanAttribute;
+import com.grookage.leia.models.attributes.ByteAttribute;
+import com.grookage.leia.models.attributes.DoubleAttribute;
+import com.grookage.leia.models.attributes.EnumAttribute;
+import com.grookage.leia.models.attributes.FloatAttribute;
+import com.grookage.leia.models.attributes.IntegerAttribute;
+import com.grookage.leia.models.attributes.LongAttribute;
+import com.grookage.leia.models.attributes.MapAttribute;
+import com.grookage.leia.models.attributes.ObjectAttribute;
+import com.grookage.leia.models.attributes.StringAttribute;
 import com.grookage.leia.models.schema.SchemaDetails;
 import com.grookage.leia.models.schema.SchemaValidationType;
 import com.grookage.leia.validator.exception.ValidationErrorCode;
 import com.grookage.leia.validator.utils.SchemaValidationUtils;
-import lombok.*;
+import lombok.SneakyThrows;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 
 class SchemaValidationUtilsTest {
 
-    @Data
-    @AllArgsConstructor
-    @NoArgsConstructor
-    @Getter
     static class ValidTestClass {
-        private Set<String> testAttribute;
+        Set<String> testAttribute;
+        TestEnum testAttribute2;
+        String testAttribute3;
+    }
+
+    enum TestEnum {
+        TEST_ENUM
     }
 
     @Test
@@ -50,11 +64,22 @@ class SchemaValidationUtilsTest {
         Assertions.assertFalse(SchemaValidationUtils.valid(schemaDetails, ValidTestClass.class));
     }
 
+    static class InvalidTestClass {
+        Set<String> testAttribute;
+    }
+
+    @Test
+    @SneakyThrows
+    void testInvalidMatchingSchema() {
+        final var schemaDetails = ResourceHelper
+                .getResource("validSchema.json", SchemaDetails.class);
+        schemaDetails.setValidationType(SchemaValidationType.MATCHING);
+        Assertions.assertNotNull(schemaDetails);
+        Assertions.assertFalse(SchemaValidationUtils.valid(schemaDetails, InvalidTestClass.class));
+    }
+
     @Test
     void testAllFields() {
-        final var arrayAttribute = new ArrayAttribute("testAttribute", true, null);
-        Assertions.assertTrue(SchemaValidationUtils.valid(Set.class, arrayAttribute));
-
         final var booleanAttribute = new BooleanAttribute("testAttribute", true, null);
         Assertions.assertTrue(SchemaValidationUtils.valid(Boolean.class, booleanAttribute));
 
@@ -76,18 +101,142 @@ class SchemaValidationUtilsTest {
         final var longAttribute = new LongAttribute("testAttribute", true, null);
         Assertions.assertTrue(SchemaValidationUtils.valid(Long.class, longAttribute));
 
-        final var mapAttribute = new MapAttribute("testAttribute", true, null);
-        Assertions.assertTrue(SchemaValidationUtils.valid(Map.class, mapAttribute));
-
-        final var objectAttribute = new ObjectAttribute("testAttribute", true, null);
-        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaDetails.class, objectAttribute));
-
         final var stringAttribute = new StringAttribute("testAttribute", true, null);
         Assertions.assertTrue(SchemaValidationUtils.valid(String.class, stringAttribute));
+
+        final var arrayAttribute = new ArrayAttribute("testAttribute", true, null, stringAttribute);
+        Assertions.assertTrue(SchemaValidationUtils.valid(Set.class, arrayAttribute));
+
+        final var mapAttribute = new MapAttribute("testAttribute", true, null, stringAttribute, stringAttribute);
+        Assertions.assertTrue(SchemaValidationUtils.valid(Map.class, mapAttribute));
+
+        final var objectAttribute = new ObjectAttribute("testAttribute", true, null, Set.of(stringAttribute));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaDetails.class, objectAttribute));
 
         Assertions.assertFalse(SchemaValidationUtils.valid(Long.class, integerAttribute));
         Assertions.assertTrue(SchemaValidationUtils.valid(Long.class, objectAttribute));
 
 
+    }
+
+    static class SetTestClass {
+        Set<String> arrayAttribute;
+    }
+
+    static class ListTestClass {
+        List<String> arrayAttribute;
+    }
+
+    static class ArrayTestClass {
+        String[] arrayAttribute;
+    }
+
+    static class RawSetTestClass {
+        Set arrayAttribute;
+    }
+
+    @Test
+    void testParametrizedArray() {
+        final var stringAttribute = new StringAttribute("stringAttribute", true, null);
+        final var arrayAttribute = new ArrayAttribute("arrayAttribute", true, null, stringAttribute);
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(arrayAttribute),
+                                                          SetTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(arrayAttribute),
+                                                          ListTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(arrayAttribute),
+                                                          ArrayTestClass.class));
+        Assertions.assertFalse(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(arrayAttribute),
+                                                           RawSetTestClass.class));
+    }
+
+    @Test
+    void testRawArray() {
+        final var arrayAttribute = new ArrayAttribute("arrayAttribute", true, null, null);
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING,
+                                                          Set.of(arrayAttribute), RawSetTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING,
+                                                          Set.of(arrayAttribute), SetTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING,
+                                                          Set.of(arrayAttribute), ListTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING,
+                                                          Set.of(arrayAttribute), ArrayTestClass.class));
+    }
+
+    static class MapTestClass {
+        Map<String, String> mapAttribute;
+    }
+
+    static class ConcurrentMapTestClass {
+        ConcurrentHashMap<String, String> mapAttribute;
+    }
+
+    static class RawMapTestClass {
+        Map mapAttribute;
+    }
+
+
+    @Test
+    void testParametrizedMap() {
+        final var keyAttribute = new StringAttribute("keyAttribute", true, null);
+        final var valueAttribute = new StringAttribute("valueAttribute", true, null);
+        final var mapAttribute = new MapAttribute("mapAttribute", true, null, keyAttribute, valueAttribute);
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(mapAttribute),
+                                                          MapTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(mapAttribute),
+                                                          ConcurrentMapTestClass.class));
+        Assertions.assertFalse(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(mapAttribute),
+                                                           RawMapTestClass.class));
+    }
+
+    @Test
+    void testRawMap() {
+        final var mapAttribute = new MapAttribute("mapAttribute", true, null, null, null);
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(mapAttribute),
+                                                          RawMapTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(mapAttribute),
+                                                          MapTestClass.class));
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(mapAttribute),
+                                                          ConcurrentMapTestClass.class));
+    }
+
+    static class ValidObjectTestClass {
+        String stringAttribute;
+        ValidNestedObjectTestClass nestedObjectAttribute;
+    }
+
+    static class ValidNestedObjectTestClass {
+        Integer integerAttribute;
+    }
+
+    static class InvalidObjectTestClass {
+        String stringAttribute;
+        InvalidNestedObjectTestClass nestedObjectAttribute;
+    }
+
+    static class InvalidNestedObjectTestClass {
+        String integerAttribute;
+    }
+
+    @Test
+    @SneakyThrows
+    void testNestedObject() {
+        final var schemaDetails = ResourceHelper
+                .getResource("validNestedSchema.json", SchemaDetails.class);
+        schemaDetails.setValidationType(SchemaValidationType.MATCHING);
+        Assertions.assertTrue(SchemaValidationUtils.valid(schemaDetails, ValidObjectTestClass.class));
+        Assertions.assertFalse(SchemaValidationUtils.valid(schemaDetails, InvalidObjectTestClass.class));
+    }
+
+    static class GenericArrayTestClass {
+        List<String>[] arrayAttribute;
+    }
+
+    @Test
+    void testGenericArrayType() {
+        final var stringAttribute = new StringAttribute("stringAttribute", true, null);
+        final var listAttribute = new ArrayAttribute("listAttribute", true, null, stringAttribute);
+        final var arrayAttribute = new ArrayAttribute("arrayAttribute", true, null, listAttribute);
+        Assertions.assertTrue(SchemaValidationUtils.valid(SchemaValidationType.MATCHING, Set.of(arrayAttribute),
+                                                          GenericArrayTestClass.class));
     }
 }